Add files via upload
This a simple program written in Java. It is my own introduction to graphical user interface.
This program prompts for a name with an input dialog and greet the user according to the time of the day.
For this project's sake.
5 -10 is morning.
11is day.
12-16 is afternoon.
17 - 20 evening
21 - 4 is Night.

This program uses JOptionPane for Dialog and calendar to get system time.